+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 version3.456
 Librairies updated
-
+fix categories gpt call js
+add more language information inside the prompt
+Add AI LLM model
 
 ------------------