+++ CHANGELOG.md
@@ -4,6 +4,10 @@ fix categories gpt call js
 fix error date (customer admin creation)
 add more language information inside the prompt
 Add AI LLM model
+fix small error
+
+
+
 
 ------------------