+++ CHANGELOG.md
@@ -5,7 +5,7 @@ fix error date (customer admin creation)
 add more language information inside the prompt
 Add AI LLM model
 fix small error
-
+Add new Ollama model